requirements.txt
streamlit
python-dotenv
tqdm
pymupdf